+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.13.4-dev
 
+* Throw a more useful error when a client is used after it has been closed.
+
 ## 0.13.3
 
 * Validate that the `method` parameter of BaseRequest is a valid "token".